refactor: shuffling dependencies
